requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-braces==1.11.0
django-cors-headers==2.1.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.6.4
idna==2.6
oauthlib==2.0.2
PyJWT==1.5.3
python-openid==2.2.5
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
